+++ app/src/main/kotlin/com/cmpt395/deferralapp/ui/theme/Theme.kt
@@ -1,6 +1,5 @@
-package com.cmpt395.ui.ui.theme
+package com.cmpt395.deferralapp.ui.theme
 
-import android.app.Activity
 import android.os.Build
 import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.material3.MaterialTheme
