+++ app/src/main/kotlin/com/cmpt395/deferralapp/ui/theme/Type.kt
@@ -1,4 +1,4 @@
-package com.cmpt395.ui.ui.theme
+package com.cmpt395.deferralapp.ui.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
